# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlucoTrack API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GlucoseReadingUnit =
  (typeof GlucoseReadingUnit)[keyof typeof GlucoseReadingUnit];

export const GlucoseReadingUnit = {
  mgdl: "mgdl",
  mmol: "mmol",
} as const;

export type GlucoseReadingContext =
  | (typeof GlucoseReadingContext)[keyof typeof GlucoseReadingContext]
  | null;

export const GlucoseReadingContext = {
  fasting: "fasting",
  before_meal: "before_meal",
  after_meal: "after_meal",
  bedtime: "bedtime",
  random: "random",
} as const;

export interface GlucoseReading {
  id: number;
  /** Blood glucose level in mg/dL */
  value: number;
  unit: GlucoseReadingUnit;
  context?: GlucoseReadingContext;
  notes?: string | null;
  recordedAt: string;
  createdAt: string;
}

export type CreateGlucoseReadingUnit =
  (typeof CreateGlucoseReadingUnit)[keyof typeof CreateGlucoseReadingUnit];

export const CreateGlucoseReadingUnit = {
  mgdl: "mgdl",
  mmol: "mmol",
} as const;

export type CreateGlucoseReadingContext =
  | (typeof CreateGlucoseReadingContext)[keyof typeof CreateGlucoseReadingContext]
  | null;

export const CreateGlucoseReadingContext = {
  fasting: "fasting",
  before_meal: "before_meal",
  after_meal: "after_meal",
  bedtime: "bedtime",
  random: "random",
} as const;

export interface CreateGlucoseReading {
  value: number;
  unit: CreateGlucoseReadingUnit;
  context?: CreateGlucoseReadingContext;
  notes?: string | null;
  recordedAt: string;
}

export interface Medication {
  id: number;
  name: string;
  dosage?: string | null;
  frequency?: string | null;
  color?: string | null;
  createdAt: string;
}

export interface CreateMedication {
  name: string;
  dosage?: string | null;
  frequency?: string | null;
  color?: string | null;
}

export interface MedicationLog {
  id: number;
  medicationId: number;
  medicationName: string;
  dosage?: string | null;
  notes?: string | null;
  takenAt: string;
  createdAt: string;
}

export interface CreateMedicationLog {
  medicationId: number;
  notes?: string | null;
  takenAt: string;
}

export type MealMealType =
  | (typeof MealMealType)[keyof typeof MealMealType]
  | null;

export const MealMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
  other: "other",
} as const;

export interface Meal {
  id: number;
  description: string;
  mealType?: MealMealType;
  carbsGrams?: number | null;
  notes?: string | null;
  eatenAt: string;
  createdAt: string;
}

export type CreateMealMealType =
  | (typeof CreateMealMealType)[keyof typeof CreateMealMealType]
  | null;

export const CreateMealMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
  other: "other",
} as const;

export interface CreateMeal {
  description: string;
  mealType?: CreateMealMealType;
  carbsGrams?: number | null;
  notes?: string | null;
  eatenAt: string;
}

export interface GlucoseStats {
  average?: number | null;
  min?: number | null;
  max?: number | null;
  lowCount: number;
  highCount: number;
  inRangeCount: number;
  totalReadings: number;
  /** Percentage of readings in target range */
  timeInRange: number;
  /** Estimated HbA1c percentage */
  hba1cEstimate?: number | null;
}

export type GetGlucoseReadingsParams = {
  period?: GetGlucoseReadingsPeriod;
};

export type GetGlucoseReadingsPeriod =
  (typeof GetGlucoseReadingsPeriod)[keyof typeof GetGlucoseReadingsPeriod];

export const GetGlucoseReadingsPeriod = {
  day: "day",
  week: "week",
  month: "month",
  year: "year",
} as const;

export type GetMedicationLogsParams = {
  period?: GetMedicationLogsPeriod;
};

export type GetMedicationLogsPeriod =
  (typeof GetMedicationLogsPeriod)[keyof typeof GetMedicationLogsPeriod];

export const GetMedicationLogsPeriod = {
  day: "day",
  week: "week",
  month: "month",
  year: "year",
} as const;

export type GetMealsParams = {
  period?: GetMealsPeriod;
};

export type GetMealsPeriod =
  (typeof GetMealsPeriod)[keyof typeof GetMealsPeriod];

export const GetMealsPeriod = {
  day: "day",
  week: "week",
  month: "month",
  year: "year",
} as const;

export type GetStatsParams = {
  period?: GetStatsPeriod;
};

export type GetStatsPeriod =
  (typeof GetStatsPeriod)[keyof typeof GetStatsPeriod];

export const GetStatsPeriod = {
  day: "day",
  week: "week",
  month: "month",
  year: "year",
} as const;