# lib/api-zod/src/generated/types/createGlucoseReading.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlucoTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGlucoseReadingContext } from "./createGlucoseReadingContext";
import type { CreateGlucoseReadingUnit } from "./createGlucoseReadingUnit";

export interface CreateGlucoseReading {
  value: number;
  unit: CreateGlucoseReadingUnit;
  context?: CreateGlucoseReadingContext;
  notes?: string | null;
  recordedAt: Date;
}