# lib/api-zod/src/generated/types/createMedication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlucoTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMedication {
  name: string;
  dosage?: string | null;
  frequency?: string | null;
  color?: string | null;
}