# lib/api-zod/src/generated/types/createMedicationLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlucoTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMedicationLog {
  medicationId: number;
  notes?: string | null;
  takenAt: Date;
}