# lib/api-zod/src/generated/types/glucoseReading.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlucoTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GlucoseReadingContext } from "./glucoseReadingContext";
import type { GlucoseReadingUnit } from "./glucoseReadingUnit";

export interface GlucoseReading {
  id: number;
  /** Blood glucose level in mg/dL */
  value: number;
  unit: GlucoseReadingUnit;
  context?: GlucoseReadingContext;
  notes?: string | null;
  recordedAt: Date;
  createdAt: Date;
}