# lib/api-zod/src/generated/types/glucoseStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlucoTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GlucoseStats {
  average?: number | null;
  min?: number | null;
  max?: number | null;
  lowCount: number;
  highCount: number;
  inRangeCount: number;
  totalReadings: number;
  /** Percentage of readings in target range */
  timeInRange: number;
  /** Estimated HbA1c percentage */
  hba1cEstimate?: number | null;
}