# lib/api-zod/src/generated/types/meal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlucoTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealMealType } from "./mealMealType";

export interface Meal {
  id: number;
  description: string;
  mealType?: MealMealType;
  carbsGrams?: number | null;
  notes?: string | null;
  eatenAt: Date;
  createdAt: Date;
}